+++ examples/actionable_observe_example.ts
@@ -1,55 +1,63 @@
 /**
- * This file is meant to be used as a scratchpad for trying out actionable observe.
- * To create a Stagehand project with best practices and configuration, run:
+ * This example shows how to use actionable observe()
  *
- * npx create-browser-app@latest my-browser-app
+ * You can use observe to get a cache-able Playwright action as JSON, then pass that JSON to act() to perform the action.
+ *
+ * This is useful for:
+ * - Previewing actions before running them
+ * - Saving actions to a file and replaying them later
+ * - Hiding sensitive information from LLMs
  */
 
-import { Stagehand } from "@/dist";
-import stagehandConfig from "@/stagehand.config";
+import { ObserveResult, Stagehand } from "@browserbasehq/stagehand";
+import StagehandConfig from "../stagehand.config";
 
 async function example() {
-  const stagehand = new Stagehand(stagehandConfig);
+  const stagehand = new Stagehand(StagehandConfig);
   await stagehand.init();
-  await stagehand.page.goto("https://www.apartments.com/san-francisco-ca/");
+  const page = stagehand.page;
+
+  await page.goto("https://www.apartments.com/san-francisco-ca/");
+
+  let observation: ObserveResult;
 
   await new Promise((resolve) => setTimeout(resolve, 3000));
-  const observations1 = await stagehand.page.observe({
+  [observation] = await page.observe({
     instruction: "find the 'all filters' button",
   });
-  await stagehand.page.act(observations1[0]);
+  await page.act(observation);
 
   await new Promise((resolve) => setTimeout(resolve, 3000));
-  const observations2 = await stagehand.page.observe({
+  [observation] = await page.observe({
     instruction: "find the '1+' button in the 'beds' section",
   });
-  await stagehand.page.act(observations2[0]);
+  await page.act(observation);
 
   await new Promise((resolve) => setTimeout(resolve, 3000));
-  const observations3 = await stagehand.page.observe({
+  [observation] = await page.observe({
     instruction: "find the 'apartments' button in the 'home type' section",
   });
-  await stagehand.page.act(observations3[0]);
+  await page.act(observation);
 
   await new Promise((resolve) => setTimeout(resolve, 3000));
-  const observations4 = await stagehand.page.observe({
+  [observation] = await page.observe({
     instruction: "find the pet policy dropdown to click on.",
   });
-  await stagehand.page.act(observations4[0]);
+  await page.act(observation);
 
   await new Promise((resolve) => setTimeout(resolve, 3000));
-  const observations5 = await stagehand.page.observe({
+  [observation] = await page.observe({
     instruction: "find the 'Dog Friendly' option to click on",
   });
-  await stagehand.page.act(observations5[0]);
+  await page.act(observation);
 
   await new Promise((resolve) => setTimeout(resolve, 3000));
-  const observations6 = await stagehand.page.observe({
+  [observation] = await page.observe({
     instruction: "find the 'see results' section",
   });
-  await stagehand.page.act(observations6[0]);
+  await page.act(observation);
 
-  const currentUrl = await stagehand.page.url();
+  const currentUrl = page.url();
   await stagehand.close();
   if (
     currentUrl.includes(
