+++ examples/example.ts
@@ -5,22 +5,23 @@
  * npx create-browser-app@latest my-browser-app
  */
 
-import { Stagehand } from "@/dist";
-import StagehandConfig from "@/stagehand.config";
+import { Stagehand } from "@browserbasehq/stagehand";
+import StagehandConfig from "../stagehand.config";
 
-async function example() {
+async function example(stagehand: Stagehand) {
+  /**
+   * Add your code here!
+   */
+  const page = stagehand.page;
+  await page.goto("https://docs.stagehand.dev");
+  await page.act("click the quickstart button");
+}
+
+(async () => {
   const stagehand = new Stagehand({
     ...StagehandConfig,
   });
   await stagehand.init();
-  /**
-   * Add your code here!
-   */
-  await stagehand.page.goto("https://docs.stagehand.dev");
-  await stagehand.page.act("click the quickstart button");
+  await example(stagehand);
   await stagehand.close();
-}
-
-(async () => {
-  await example();
 })();
