+++ package.json
@@ -6,21 +6,7 @@
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "scripts": {
-    "2048": "pnpm run build && tsx examples/2048.ts",
-    "popup": "pnpm run build && tsx examples/popup.ts",
-    "cua": "pnpm run build && tsx examples/cua-example.ts",
-    "operator": "pnpm run build && tsx examples/operator-example.ts",
-    "example": "pnpm run build && tsx examples/example.ts",
-    "langchain": "pnpm run build && tsx examples/langchain.ts",
-    "debug-url": "pnpm run build && tsx examples/debugUrl.ts",
-    "external-client": "pnpm run build && tsx examples/external_client.ts",
-    "instructions": "pnpm run build && tsx examples/instructions.ts",
-    "ai-sdk-client": "pnpm run build && tsx examples/ai_sdk_example.ts",
-    "actionable_observe_example": "pnpm run build && tsx examples/actionable_observe_example.ts",
-    "form-filling-sensible-cerebras": "pnpm run build && tsx examples/form_filling_sensible_cerebras.ts",
-    "form-filling-sensible": "pnpm run build && tsx examples/form_filling_sensible.ts",
-    "google-enter": "pnpm run build && tsx examples/google_enter.ts",
-    "try-wordle": "pnpm run build && tsx examples/try_wordle.ts",
+	"example": "pnpm --filter @browserbasehq/stagehand-examples run start",
     "format": "prettier --write .",
     "prettier": "prettier --check .",
     "prettier:fix": "prettier --write .",
@@ -37,7 +23,21 @@
     "prepare": "pnpm run build",
     "lint": "pnpm run prettier && pnpm run eslint",
     "release": "pnpm run build && changeset publish",
-    "release-canary": "pnpm run build && changeset version --snapshot && changeset publish --tag alpha"
+    "release-canary": "pnpm run build && changeset version --snapshot && changeset publish --tag alpha",
+    "2048": "pnpm --filter @browserbasehq/stagehand-examples run start 2048",
+    "popup": "pnpm --filter @browserbasehq/stagehand-examples run start popup",
+    "cua": "pnpm --filter @browserbasehq/stagehand-examples run start cua",
+    "operator": "pnpm --filter @browserbasehq/stagehand-examples run start operator",
+    "langchain": "pnpm --filter @browserbasehq/stagehand-examples run start custom_client_langchain",
+    "external-client": "pnpm --filter @browserbasehq/stagehand-examples run start custom_client_openai",
+    "ai-sdk-client": "pnpm --filter @browserbasehq/stagehand-examples run start custom_client_aisdk",
+    "debug-url": "pnpm --filter @browserbasehq/stagehand-examples run start debug-url",
+    "instructions": "pnpm --filter @browserbasehq/stagehand-examples run start instructions",
+    "actionable_observe_example": "pnpm --filter @browserbasehq/stagehand-examples run start actionable_observe_example",
+    "form-filling-sensible-cerebras": "pnpm --filter @browserbasehq/stagehand-examples run start form-filling-sensible-cerebras",
+    "form-filling-sensible": "pnpm --filter @browserbasehq/stagehand-examples run start form-filling-sensible",
+    "google-enter": "pnpm --filter @browserbasehq/stagehand-examples run start google-enter",
+    "try-wordle": "pnpm --filter @browserbasehq/stagehand-examples run start try-wordle"
   },
   "files": [
     "dist/**"
